requirements.txt
pynput>=1.7.6
mss>=9.0.1
numpy>=1.24.0
Pillow>=10.0.0
